Total management worker event valar

+++ database/migrations/2023_05_22_070929_create_directrecruitment_onboarding_countries_table.php
@@ -46,6 +46,8 @@ return new class extends Migration
      */
     public function down(): void
     {
+      if (DB::getDriverName() !== 'sqlite') {
         Schema::dropIfExists('directrecruitment_onboarding_countries');
+      }
     }
 };
